requirements.txt
torchvision==0.10.0
torch==1.9.0
Pillow==8.2.0
PyYAML==5.4.1
tensorflow==2.3.0
joblib==0.17.0
numpy==1.18.5
matplotlib==3.3.3
opencv-python==4.5.2.54
pandas==1.2.5
seaborn==0.11.1
yaml~=0.2.5
imagehash~=4.2.0
face-detection==0.2.2

